* moved private elements into **internal** (so as to hide properly); * dependencies; * GitHub Actions enhancements; * boilerplate;
