+++ .github/workflows/go.yml
@@ -5,21 +5,29 @@ name: Go
 
 on:
   push:
-    branches: [ "master" ]
+    branches: [ "master", "dev", "gha" ]
   pull_request:
-    branches: [ "master" ]
+    branches: [ "master", "dev", "gha" ]
 
 jobs:
 
   build:
     runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        go-version: [ '1.21', '1.22', '1.23', '1.24' ]
+
     steps:
     - uses: actions/checkout@v4
 
-    - name: Set up Go
+    - name: Set up Go ${{ matrix.go-version }}
       uses: actions/setup-go@v4
       with:
-        go-version: '1.24'
+        go-version: '${{ matrix.go-version }}'
+
+    - name: Display Go version
+      run: go version
 
     - name: Build
       run: ./build.sh
