+++ internal/_std_strings/find.go
