+++ version_test.go
@@ -11,7 +11,7 @@ const (
 	Expected_VersionMajor uint16 = 0
 	Expected_VersionMinor uint16 = 7
 	Expected_VersionPatch uint16 = 0
-	Expected_VersionAB    uint16 = 0x4001
+	Expected_VersionAB    uint16 = 0x8001
 )
 
 func Test_Version_Elements(t *testing.T) {
@@ -22,9 +22,9 @@ func Test_Version_Elements(t *testing.T) {
 }
 
 func Test_Version(t *testing.T) {
-	require.Equal(t, uint64(0x0000_0007_0000_4001), angols.Version)
+	require.Equal(t, uint64(0x0000_0007_0000_8001), angols.Version)
 }
 
 func Test_Version_String(t *testing.T) {
-	require.Equal(t, "0.7.0-alpha1", angols.VersionString())
+	require.Equal(t, "0.7.0-beta1", angols.VersionString())
 }
